Add notify on true code

+++ restic-backup.sh
@@ -48,8 +48,10 @@ echo "Uploading data to Amazon S3"
 OUTPUT=$(/root/OT-Smoothbrain-Backup/restic backup /root/backup/.origintrail_noderc /root/backup/* 2>&1)
 
 if [ $? -eq 0 ]; then
-  /root/OT-Settings/data/send.sh "Backup SUCCESSFUL:${N1}$OUTPUT"
-  rm -rf /root/backup/* /root/backup/.origintrail_noderc
+  if $SMOOTHBRAIN_NOTIFY_ON_SUCCESS == "true"; then
+    /root/OT-Settings/data/send.sh "Backup SUCCESSFUL:${N1}$OUTPUT"
+    rm -rf /root/backup/* /root/backup/.origintrail_noderc
+  fi
 else
   /root/OT-Settings/data/send.sh "Uploading backup to S3 FAILED:${N1}$OUTPUT"
   exit 1
